refactor: SectorRotation 별도 파일 분리 (HomeDashboard 901→819줄)
- SectorRotation.jsx 신규 생성 (82줄)
- HomeDashboard.jsx에서 인라인 정의 제거 후 import 교체
